Create OutlookCalendarEvents.py
This script retrieves your calendar events from Office365/Outlook. 
It requires you to register your app in the Azuer Portal, so it will generate the app id, secret, etc for this script to work. 
It will give you a URL so you can generate a code that will grant you access to your calendar events